+++ pyreadstat/conditional_includes.h
@@ -20,7 +20,7 @@
 
     //int open(const char *path, int oflag, int mode);
     //int close(int fd);
-    int _wsopen(const wchar_t *filename, int oflag, int shflag, int pmode){};
+    int _wsopen(const wchar_t *filename, int oflag, int shflag, int pmode){ return 0; };
     int _O_RDONLY;
     int _O_BINARY;
     int _O_WRONLY;
@@ -31,6 +31,6 @@
     int _SH_DENYRD;
     int _SH_DENYNO;
     void assign_fd(void *io_ctx, int fd){};
-    int _close(int fd){};
+    int _close(int fd){ return 0; };
     
 #endif