+++ pyreadstat/pyreadstat.pyx
@@ -505,7 +505,9 @@ def set_catalog_to_sas(sas_dataframe, sas_metadata, catalog_metadata, formats_as
 
 # Write API
 
-def write_sav(df, str dst_path, str file_label="", list column_labels=None, compress=False, str note=None):
+def write_sav(df, str dst_path, str file_label="", list column_labels=None, compress=False, str note=None,
+                dict variable_value_labels=None, dict missing_ranges=None, dict variable_display_width=None,
+                dict variable_measure=None):
     """
     Writes a pandas data frame to a SPSS sav or zsav file.
 
@@ -524,16 +526,39 @@ def write_sav(df, str dst_path, str file_label="", list column_labels=None, comp
         if true a zsav will be written, by default False, a sav is written
     note : str, optional
         a note to add to the file
-
+    variable_value_labels : dict, optional
+        value labels, a dictionary with key variable name and value a dictionary with key values and
+        values labels. Variable names must match variable names in the dataframe otherwise will be
+        ignored. Value types must match the type of the column in the dataframe.
+    missing_ranges : dict, optional
+        user defined missing values. Must be a dictionary with keys as variable names matching variable
+        names in the dataframe. The values must be a list. Each element in that list can either be
+        either a discrete numeric or string value (max 3 per variable) or a dictionary with keys 'hi' and 'lo' to
+        indicate the upper and lower range for numeric values (max 1 range value + 1 discrete value per
+        variable). hi and lo may also be the same value in which case it will be interpreted as a discrete
+        missing value.
+        For this to be effective, values in the dataframe must be the same as reported here and not NaN.
+    variable_display_width : dict, optional
+        set the display width for variables. Must be a dictonary with keys being variable names and
+        values being integers.
+    variable_measure: dict, optional
+        sets the measure type for a variable. Must be a dictionary with keys being variable names and
+        values being strings one of "nominal", "ordinal", "scale" or "unknown" (default).
     """
 
     cdef int file_format_version = 2
     if compress:
         file_format_version = 3
     cdef table_name = ""
-    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_SAV, file_label, column_labels, file_format_version, note, table_name)
+    cdef dict missing_user_values = None
+    cdef dict variable_alignment = None
+    
+    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_SAV, file_label, column_labels, 
+        file_format_version, note, table_name, variable_value_labels, missing_ranges, missing_user_values,
+        variable_alignment, variable_display_width, variable_measure)
 
-def write_dta(df, str dst_path, str file_label="", list column_labels=None, int version=15):
+def write_dta(df, str dst_path, str file_label="", list column_labels=None, int version=15, 
+            dict variable_value_labels=None, dict missing_user_values=None):
     """
     Writes a pandas data frame to a STATA dta file
 
@@ -550,7 +575,14 @@ def write_dta(df, str dst_path, str file_label="", list column_labels=None, int 
         labels must be represented by None.
     version : int, optional
         dta file version, supported from 8 to 15, default is 15
-
+    variable_value_labels : dict, optional
+        value labels, a dictionary with key variable name and value a dictionary with key values and
+        values labels. Variable names must match variable names in the dataframe otherwise will be
+        ignored. Value types must match the type of the column in the dataframe.
+    missing_user_values : dict, optional
+        user defined missing values for numeric variables. Must be a dictionary with keys being variable
+        names and values being a list of missing values. Missing values must be a single character
+        between a and z.
     """
 
     if version == 15:
@@ -569,8 +601,15 @@ def write_dta(df, str dst_path, str file_label="", list column_labels=None, int 
         raise Exception("Version not supported")
 
     cdef str note = ""
-    cdef table_name = ""
-    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_DTA, file_label, column_labels, file_format_version, note, table_name)
+    cdef str table_name = ""
+    cdef dict missing_ranges = None
+    cdef dict variable_alignment = None
+    cdef dict variable_display_width = None
+    cdef dict variable_measure = None
+
+    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_DTA, file_label, column_labels, file_format_version,
+     note, table_name, variable_value_labels, missing_ranges, missing_user_values, variable_alignment,
+     variable_display_width, variable_measure)
 
 def write_xport(df, str dst_path, str file_label="", list column_labels=None, str table_name=None):
     """
@@ -595,6 +634,13 @@ def write_xport(df, str dst_path, str file_label="", list column_labels=None, st
 
     # atm version 5 and 8 are supported by readstat but only 5 can be later be read by SAS
     cdef int file_format_version = 5
-
+    cdef dict variable_value_labels=None
     cdef str note = ""
-    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_XPORT, file_label, column_labels, file_format_version, note, table_name)
+    cdef dict missing_ranges = None
+    cdef dict missing_user_values = None
+    cdef dict variable_alignment = None
+    cdef dict variable_display_width = None
+    cdef dict variable_measure = None
+    run_write(df, dst_path, _readstat_writer.FILE_FORMAT_XPORT, file_label, column_labels, 
+        file_format_version, note, table_name, variable_value_labels, missing_ranges,missing_user_values,
+        variable_alignment,variable_display_width, variable_measure)