+++ tests/test_basic.py
@@ -407,8 +407,10 @@ class TestBasic(unittest.TestCase):
         df_sas, meta = pyreadstat.read_sas7bdat(sas_file, user_missing=True)
         df_csv = pd.read_csv(formatted_csv)
         self.assertTrue(df_sas.equals(df_csv))
-        self.assertTrue(meta.missing_user_values == ['A', 'B', 'C', 'X', 'Y', 'Z', '_'])
-        
+        missing_user_values = {'var1':['A'],'var2': ['B'], 'var3':['C'], 'var4':['X'], 'var5':['Y'], 
+        'var6':['Z'], 'var7':['_']}
+        self.assertDictEqual(meta.missing_user_values, missing_user_values)
+
         df_sas, meta = pyreadstat.read_sas7bdat(sas_file,
                             catalog_file=cat_file, user_missing=True,
                             formats_as_category=False)
@@ -428,7 +430,8 @@ class TestBasic(unittest.TestCase):
         df_sas, meta = pyreadstat.read_dta(dta_file, user_missing=True)
         df_csv = pd.read_csv(formatted_csv)
         self.assertTrue(df_sas.equals(df_csv))
-        self.assertTrue(meta.missing_user_values == ['a', 'b', 'c', 'x', 'y', 'z'])
+        missing_user_values = {'var1':['a'],'var2': ['b'], 'var3':['c'], 'var4':['x'], 'var5':['y'], 'var6':['z']}
+        self.assertDictEqual(meta.missing_user_values, missing_user_values)
         
         df_sas, meta = pyreadstat.read_dta(dta_file,
                             apply_value_formats=True, user_missing=True,
@@ -472,40 +475,46 @@ class TestBasic(unittest.TestCase):
 
     def test_sav_write_basic(self):
 
-        #if sys.version_info[0] < 3:
-        #    return
-
         file_label = "basic write"
         file_note = "These are some notes"
         col_labels = ["mychar label","mynum label", "mydate label", "dtime label", None, "myord label", "mytime label"]
+        variable_value_labels = {'mylabl': {1.0: 'Male', 2.0: 'Female'}, 'myord': {1.0: 'low', 2.0: 'medium', 3.0: 'high'}}
+        missing_ranges = {'mychar':['a'], 'myord': [{'hi':2, 'lo':1}]}
+        #variable_alignment = {'mychar':"center", 'myord':"right"}
+        variable_display_width = {'mychar':20}
+        variable_measure = {"mychar": "nominal"}
         path = os.path.join(self.write_folder, "basic_write.sav")
-        pyreadstat.write_sav(self.df_pandas, path, file_label=file_label, column_labels=col_labels, note=file_note)
-        df, meta = pyreadstat.read_sav(path)
+        pyreadstat.write_sav(self.df_pandas, path, file_label=file_label, column_labels=col_labels, note=file_note, 
+            variable_value_labels=variable_value_labels, missing_ranges=missing_ranges, variable_display_width=variable_display_width,
+            variable_measure=variable_measure) #, variable_alignment=variable_alignment)
+        df, meta = pyreadstat.read_sav(path, user_missing=True)
         self.assertTrue(df.equals(self.df_pandas))
         self.assertEqual(meta.file_label, file_label)
         self.assertListEqual(meta.column_labels, col_labels)
         self.assertEqual(meta.notes[0], file_note)
+        self.assertDictEqual(meta.variable_value_labels, variable_value_labels)
+        self.assertEqual(meta.variable_display_width['mychar'], variable_display_width['mychar'])
+        #self.assertDictEqual(meta.variable_alignment, variable_alignment)
+        self.assertEqual(meta.variable_measure["mychar"], variable_measure["mychar"])
 
     def test_zsav_write_basic(self):
 
-        #if sys.version_info[0] < 3:
-        #    return
-
         file_label = "basic write"
         file_note = "These are some notes"
         col_labels = ["mychar label","mynum label", "mydate label", "dtime label", None, "myord label", "mytime label"]
+        variable_value_labels = {'mylabl': {1.0: 'Male', 2.0: 'Female'}, 'myord': {1.0: 'low', 2.0: 'medium', 3.0: 'high'}}
+        missing_ranges = {'mychar':['a'], 'myord': [{'hi':2, 'lo':1}]}
         path = os.path.join(self.write_folder, "basic_write.zsav")
-        pyreadstat.write_sav(self.df_pandas, path, file_label=file_label, column_labels=col_labels, compress=True, note=file_note)
-        df, meta = pyreadstat.read_sav(path)
+        pyreadstat.write_sav(self.df_pandas, path, file_label=file_label, column_labels=col_labels, compress=True, note=file_note,
+                     variable_value_labels=variable_value_labels, missing_ranges=missing_ranges)
+        df, meta = pyreadstat.read_sav(path, user_missing=True)
         self.assertTrue(df.equals(self.df_pandas))
         self.assertEqual(meta.file_label, file_label)
         self.assertListEqual(meta.column_labels, col_labels)
         self.assertEqual(meta.notes[0], file_note)
+        self.assertDictEqual(meta.variable_value_labels, variable_value_labels)
 
     def test_dta_write_basic(self):
-
-        #if sys.version_info[0] < 3:
-        #    return
 
         df_pandas = self.df_pandas.copy()
         df_pandas["myord"] = df_pandas["myord"].astype(np.int32)
@@ -513,8 +522,9 @@ class TestBasic(unittest.TestCase):
 
         file_label = "basic write"
         col_labels = ["mychar label","mynum label", "mydate label", "dtime label", None, "myord label", "mytime label"]
+        variable_value_labels = {'mylabl': {1: 'Male', 2: 'Female'}, 'myord': {1: 'low', 2: 'medium', 3: 'high'}}
         path = os.path.join(self.write_folder, "basic_write.dta")
-        pyreadstat.write_dta(df_pandas, path, file_label=file_label, column_labels=col_labels, version=12)
+        pyreadstat.write_dta(df_pandas, path, file_label=file_label, column_labels=col_labels, version=12, variable_value_labels=variable_value_labels)
         df, meta = pyreadstat.read_dta(path)
 
         df_pandas["myord"] = df_pandas["myord"].astype(np.int64)
@@ -523,11 +533,30 @@ class TestBasic(unittest.TestCase):
         self.assertTrue(df.equals(df_pandas))
         self.assertEqual(meta.file_label, file_label)
         self.assertListEqual(meta.column_labels, col_labels)
+        self.assertDictEqual(meta.variable_value_labels, variable_value_labels)
+
+    def test_dta_write_user_missing(self):
+
+
+        #formatted_csv = os.path.join(self.missing_data_folder, "missing_dta_formatted.csv")
+        #labeled_csv = os.path.join(self.missing_data_folder, "missing_dta_labeled.csv")
+        
+        df_csv = pd.DataFrame([[3,"a"],["a","b"]], columns=["Var1", "Var2"])
+        df_csv2 = pd.DataFrame([[3,"a"],["labeled","b"]], columns=["Var1", "Var2"])
+
+        missing_user_values = {'Var1': ['a']}
+        variable_value_labels = {'Var1':{'a':'labeled'}}
+        path = os.path.join(self.write_folder, "user_missing_write.dta")
+        pyreadstat.write_dta(df_csv, path, version=12, missing_user_values=missing_user_values, variable_value_labels=variable_value_labels)
+        
+        df_dta, meta = pyreadstat.read_dta(path, user_missing=True)
+        self.assertTrue(df_csv.equals(df_dta))
+        self.assertDictEqual(meta.missing_user_values, missing_user_values)
+        
+        df_dta2, meta2 = pyreadstat.read_dta(path, user_missing=True, apply_value_formats=True, formats_as_category=False)
+        self.assertTrue(df_csv2.equals(df_dta2))
 
     def test_xport_write_basic(self):
-
-        #if sys.version_info[0] < 3:
-        #    return
 
         file_label = "basic write"
         table_name = "TEST"